requirements.txt
Flask-JWT==0.3.2
Flask-Migrate==2.0.3
flask-restplus==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
itsdangerous==0.24
psycopg2==2.7.1
